==> R/02_process_data.Rmd <==
---
title: "Process raw data"
author: "shelley herbrich"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
editor_options: 
  chunk_output_type: console
---

# Overview

The 'raw' dataset from the original Ng et al. manuscript contains gene expression data from Illumina HumanHT-12 V4.0 expression beadchip arrays, but they also have associated metadata. For review, we briefly describe the data matrix and corresponding annotation dataframes.

# Libraries

```{r load_libs, message=FALSE}

library(here)
library(readr)
library(GEOquery)
library(illuminaHumanv4.db)

```

# Ng et al

## Text Description

In late 2016, [Ng et al](https://www.nature.com/articles/nature20598) produced the first publicly available gene expression dataset functionally characterizing acute myeloid leukemia (AML) stem cells (LSCs). To date, existing datasets distinguish LSCs immunophenotypically. However, we know that the functional definition of LSCs, which is the ability to engraft in sublethally irradiated mice, may be more clinically relevant. Therefore, Ng et al sorted patient samples by CD34/CD38 +/- status into 4 groups and tested engraftment in immunocompromised mice. LSC+ samples were defined as populations that engrafted in mice.

As described in the methods: 
> RNA was extracted using Qiagen RNeasy mini kits (catalogue 74106) and was subjected to GE analysis using Illumina HumanHT-12 v4 microarrays to investigate ~ 47,000 targets corresponding to ~ 30,000 genes. The resultant fluorescence intensity profiles were subjected to variance stabilization and robust spline normalization using the lumi 2.16.0 R package31. All data was put into the log base-2 scale. 

The normalized gene expression data was downloaded from the GEO SuperSeries under accession number GSE76008 (08/24/2017 version) (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc= GSE76008). Sample identifier information was obtained from Stanley Ng.

## File Examination

```{r load_ng_as_matrix, warning=FALSE}
temp <- 
  read_delim(here("data", "GSE76009-GPL10558_series_matrix.txt.gz"), delim = "\t", skip = 56)

#############################
# gene expression stored in
# rows not beginning with '!'
#############################
dat <- 
  temp[which(sapply(temp[,1], function(x) substr(x,1,1)) != '!'),]
sampleID <- as.character(dat[1,-1])
probeID <- (dat[-1,1])
dat <- data.matrix(dat[-1,-1])
rownames(dat) <- probeID$`!Sample_title`
colnames(dat) <- sampleID

#############################
# for annotation, let's start
# by taking all rows with '!'
#############################

sample_annot <- 
  temp[which(sapply(temp[,1], function(x) substr(x,1,1)) == '!'),]
sample_annot <- sample_annot[which(sample_annot$'!Sample_title' %in% c('!Sample_geo_accession','!Sample_characteristics_ch1','!Sample_description')),]
sample_annot <- t(sample_annot[,-1])
colnames(sample_annot) <- c("sampleID",'cell_fraction','lsc_status','batch')
sample_annot <- data.frame(sample_annot)
rownames(sample_annot) <- sample_annot$batch

###########################
# grab additional sample 
# annotation- patient ID
############################

additional_annot <- read.csv(here("data", "GSE76008_patientID.csv"), header = TRUE, row.names = 1)
sample_annot <- data.frame(sample_annot, patientID= additional_annot[rownames(sample_annot),1])
rownames(sample_annot) <- sample_annot$sampleID
sample_annot <- sample_annot[colnames(dat),]
all(rownames(sample_annot)==colnames(dat))

## clean 
sample_annot$cell_fraction <- sapply(as.character(sample_annot$cell_fraction), function(x) strsplit(x, ': ')[[1]][2])
sample_annot$lsc_status <- sapply(as.character(sample_annot$lsc_status), function(x) strsplit(x, ': ')[[1]][2])

###########################
# probe annotation
############################

# Get the probe identifiers that are mapped to a gene symbol
mapped_probes <- mappedkeys(illuminaHumanv4SYMBOL)
# Convert to a list
probe_to_gene <- as.list(illuminaHumanv4SYMBOL[intersect(mapped_probes, rownames(dat))])

dat[1:5, 1:5]
sample_annot[1:5, 1:5]

## Clean up

rm(temp, sampleID, probeID, additional_annot, mapped_probes)

## Save the results

save(dat, sample_annot, probe_to_gene,
     file = here("results", "ng_data.RData"))
```
